requirements.txt
# ── UI ───────────────────────────────────────────────────────────
PyQt6>=6.6.0

# ── Audio capture + transcription ────────────────────────────────
sounddevice>=0.4.6
SpeechRecognition>=3.10.0
numpy>=1.26.0

# ── AI API SDKs ───────────────────────────────────────────────────
anthropic>=0.25.0            # Claude  →  ANTHROPIC_API_KEY
google-genai                 # Gemini  →  GOOGLE_API_KEY (New SDK)
groq>=0.9.0                  # Groq    →  GROQ_API_KEY
httpx>=0.27.0                # OpenRouter (raw HTTP)  →  OPENROUTER_API_KEY

# ── Image conversion (required for Google Gemini vision) ─────────
Pillow>=10.0.0
# ── Environment ──────────────────────────────────────────────────
python-dotenv
